Separate display and configuration logic into separate plugins
